+++ changelog.md
@@ -24,6 +24,7 @@
 * `keyToPrivateEcAddress` renamed `seedToPrivateEcAddress`.
 * `add` throws an Error when trying to create a chain that already exists.
 * All the `add*` functions validate by default that the paying EC address holds enough funds to pay for the commits. Can be bypassed by setting the option `skipFundValidation` to `true`.
+* Added a check before sending a transaction that all EC ouputs have a factoshi amount above the minimum amount required to get at least 1 EC. It can be bypassed using the `force` option of `sendTransaction` method.
 
 ### Removed